Move VDB functions from Cycles to intern/openvdb.
So they can be used on Blender side. Changed variable names and types to
be more consistent between Cycles and intern/openvdb.

+++ intern/cycles/util/util_sparse_grid.h
@@ -71,7 +71,7 @@ const inline int compute_index(const size_t x, const size_t y, const size_t z,
 	return compute_index(x, y, z, resolution.x, resolution.y, resolution.z);
 }
 
-const inline size_t get_tile_res(const size_t res)
+const inline int get_tile_res(const size_t res)
 {
 	return (res / TILE_SIZE) + (res % TILE_SIZE != 0);
 }
